move variables to independent class (rough, untested) [skip CI]
